voter_pipeline/requirements.txt
pymysql
python-dotenv
requests
openpyxl
playwright
